Refactor sidebar (create alternate authenticated version with middle buttons)
